Added ServerList message for exchanging server
addresses when server joins or rejoins.
New server only need to know its own listen
port, and if it is not the first one,
then the 2nd argument is the existing server's
port.

Server can be restarted and it requests connect
addresses when they send the first Hello-message
to it.